refactor: move core utilities
